Add files via upload
The Heart Failure Clinical Records dataset is a machine learning dataset that contains information about patients diagnosed with heart failure. The dataset includes 299 samples, each representing a patient, and 13 features.
The dataset is commonly used for binary classification problems, with the goal of predicting whether a patient is at risk of death from heart failure or not. The dataset is often used to explore the relationship between various features and the risk of death from heart failure, such as the correlation between age and risk of death, or the effect of various medical conditions on the risk of death. The dataset is also used to evaluate the performance of machine learning algorithms for binary classification tasks in the medical field.